Workaround iOS 10.x layout margin bug (Fix #2)
Setting a custom layout margin in Interface Builder on a subview does
not always take effect when the view is first loaded. The default
margin sometimes applies until the layout is redone (e.g. after
rotation).

This problem only affects margins set in the storyboard and seems to
be fixed in iOS 11.

Setting `preservesSuperviewLayoutMargins` on the view seems to be
sufficient to trigger the custom layout margin. If that is not
a convenient workaround you can fall back to setting the layout
margin in code (make sure to leave the storyboard with a default
margin).