Workaround for layout margin bug (iOS 9/10)
In iOS 9 and iOS 10 there is a known bug which can prevent the
layout margins from being applied when the view is first loaded.
It only seems to impact the margins of direct subviews of the root
view. The subview appears with the default (8 point) margin until
the layout is refreshed (for example, by rotating the device).

The bug does not appear if you are constraining the subview to the
margins of the root view and are using a non-default margin for the
root view (but note that the non-default root view margin has no
effect on the actual margin for iOS 10 and earlier).

The problem was fixed in iOS 11 and later.

Possible workarounds include:

* Apply the margin in code (make sure to leave the storyboard
with the default margin).
* Select the "Preserves Superview Margins" option for the subview
in the storyboard (the red view in this example).

I applied the second of these workarounds in this project.